Reject and remove the unchecked aggregate kernels; record the result (Issue #510)
The acceptance gate was not met, so the prototype is reverted and this commit
ships the measured record instead of an optimisation. No library source file
changes relative to the base commit.

At the production-realistic ~10% aggregate frequency the prototype measured
-1.3% to -4.8% end-to-end against a 19-40% session noise floor; the 15-38%
gains appear only at 50-100% aggregate frequency, which the gate excludes as
synthetic-only. An isolated A/B against safe attribution controls showed the
little that was there comes from slice iteration, not bounds-check removal.

The sweep also surfaced a much larger lever: one aggregate neuron costs the
whole creature the record-interleaved scoring fast path (35.8ms -> 85.9ms per
4,096-record shard, 2.4x). Filed as #514.

Kept: the aggregate-frequency bench (the only committed fixture with aggregate
neurons in a production-sized creature, and what #514 reproduces from), its
fixture helpers with tests, and the research record.

Vibe-Coder-Run-Id: vibe-msfll04t-81dbae
